test-web-mobile/appium-python/requirements.txt
Appium-Python-Client==0.24
py==1.4.34
pytest==3.2.0
selenium==3.3.1
